requirements.txt
Django>=4.2,<5.0
django-ninja>=1.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
requests>=2.31.0
openai>=1.0.0
python-dateutil>=2.8.2
